Print entryDropdown.value to the console to see its value.
Since entryDropdown.value is in a static context (outside of an event listener), it only shows the value at the moment the code runs. This means it won't automatically update as the user interacts with the dropdown.

It will capture the initial value (in this case, "breakfast") and won't reflect any changes the user makes afterward.